% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_ul.R
\name{above_ul}
\alias{above_ul}
\title{Calculate percent of population at risk of adverse effects}
\usage{
above_ul(ul, shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL, plot = F)
}
\arguments{
\item{ul}{Tolerable Upper Limit (UL)}

\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Boolean (TRUE/FALSE) indicating whether to plot the distributions relative to the EAR}
}
\value{
The percent of a population at risk of adverse effects
}
\description{
This function calculates the percent of a population whose habitual intake is above the Tolerable Upper Limit (UL) and are therefore at risk of adverse effects.
}
\examples{
above_ul(ul=11, shape=9.5, rate=1.3, plot=T)
above_ul(ul=11, meanlog=1.9, sdlog=0.3, plot=T)
above_ul(ul=13, meanlog=1.9, sdlog=0.3, plot=T)
}
