% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate the coefficient of variation (CV) of a habitual intake distribution}
\usage{
cv(shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\value{
The coefficient of variation (CV) of the distribution
}
\description{
This function calculates the coefficient of variation (CV) of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
cv(shape=1, rate=0.5)
}
