% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dists_full}
\alias{dists_full}
\title{Subnational habitual nutrient intake distributions}
\format{
A data frame with the following attributes::
\describe{
  \item{continent}{Continent}
  \item{country}{Country name}
  \item{iso3}{ISO3 code for country}
  \item{hdi_category}{Human Development Index (HDI) category}
  \item{hdi}{Human Development Index (HDI) category}
  \item{iron_type}{Iron bioavailability}
  \item{zinc_type}{Zin bioavailability}
  \item{nutrient_type}{Type of nutrient (i.e., macronutrient, vitamin, element)}
  \item{nutrient}{Nutrient name}
  \item{nutrient_units}{Nutrient name and units}
  \item{sex}{Sex (i.e., males, females, both)}
  \item{age_group}{Age group (5-yr intervals)}
  \item{sex_ear}{Sex used to assign EAR value}
  \item{age_group_ear}{Age group used to assign EAR value}
  \item{ear_units}{EAR units}
  \item{ear_cv}{Recommended coefficient of variation (CV) for the EAR value}
  \item{ear}{EAR value}
  \item{sex_ar}{Sex used to assign AR value}
  \item{age_group_ar}{Age group used to assign AR value}
  \item{ear_units}{AR units}
  \item{ar}{AR value}
  \item{g_shape}{Shape parameter for a gamma-distributed intake distribution}
  \item{g_rate}{Rate parameter for a gamma-distributed intake distribution}
  \item{g_ks}{K-S goodness-of-fit for a gamma-distributed intake distribution}
  \item{g_mu}{Mean of a gamma-distributed intake distribution}
  \item{g_cv}{Coefficient of variation (CV) of a gamma-distributed intake distribution}
  \item{g_var}{Variance a gamma-distributed intake distribution}
  \item{g_skew}{Skewness of a gamma-distributed intake distribution}
  \item{g_kurt}{Kurtosis of a gamma-distributed intake distribution}
  \item{g_sev_ear}{Summary exposure value (SEV) of a gamma-distributed intake distribution using EAR}
  \item{g_sev_ar}{Summary exposure value (SEV) of a gamma-distributed intake distribution using AR}
  \item{g_above_ul}{Proportion of population above the upper limit with the gamma distribution}
  \item{ln_meanlog}{Meanlog parameter for a lognormally-distributed intake distribution}
  \item{ln_sdlog}{Sdlog parameter for a lognormally-distributed intake distribution}
  \item{ln_ks}{K-S goodness-of-fit for a lognormally-distributed intake distribution}
  \item{ln_mu}{Mean of a lognormally-distributed intake distribution}
  \item{ln_cv}{Coefficient of variation (CV) of a lognormally-distributed intake distribution}
  \item{ln_var}{Variance a lognormally-distributed intake distribution}
  \item{ln_skew}{Skewness of a lognormally-distributed intake distribution}
  \item{ln_kurt}{Kurtosis of a lognormally-distributed intake distribution}
  \item{ln_sev_ear}{Summary exposure value (SEV) of a lognormally-distributed intake distribution using EAR}
  \item{ln_sev_ar}{Summary exposure value (SEV) of a lognormally-distributed intake distribution using AR}
  \item{ln_above_ul}{Proportion of population above the upper limit with the lognormal distribution}
  \item{best_dist}{Best distribution based on the K-S goodness-of-fit statistic}
  \item{mu}{Mean of the best intake distribution}
  \item{cv}{Coefficient of variation (CV) of the best intake distribution}
  \item{var}{Variance the best intake distribution}
  \item{skew}{Skewness of the best intake distribution}
  \item{kurt}{Kurtosis of the best intake distribution}
  \item{above_ul}{Proportion of population above the upper limit}
  \item{sev_ear}{Summary exposure value (SEV) of the best intake distribution using EAR}
  \item{sev_ar}{Summary exposure value (SEV) of the best intake distribution using AR}
  \item{ear_use}{EAR or AR}
  \item{sev}{Summary exposure value (SEV) of the best intake distribution using EAR}
  \item{cutpoint_sev_ear}{SEV calculated using cutpoint method and EAR}
  \item{cutpoint_sev_ar}{SEV calculated using cutpoint method and AR}
  \item{cutpoint_sev}{SEV calculated using cutpoint method}
}
}
\source{
Passarelli S, Free CM, Shepon A, Lee C, Moursi M, Cao L, Li Y, Crispim S, Schmidhuber J, Bromage S, Beal T, Golden CD (in prep) Global modeling of subnational habitual nutrient intake distributions. Near submission.
}
\usage{
dists_full
}
\description{
This dataset contains the subnational habitual nutrient intake distributions estimated by Passarelli et al. (in prep).
}
\keyword{datasets}
