% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dists.R
\name{generate_dists}
\alias{generate_dists}
\title{Generate subnational nutrient intake distributions}
\usage{
generate_dists(dists)
}
\arguments{
\item{dists}{A data frame of nutrient distributions}
}
\value{
A dataframe
}
\description{
This function generates data for plotting subnational nutrient intake distributions. It is provided a dataframe generated from the get_dists() function.
}
