% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dists.R
\name{get_dists}
\alias{get_dists}
\title{Get subnational nutrient intake distributions}
\usage{
get_dists(nutrients = NULL, isos = NULL, sexes = NULL, ages = NULL)
}
\arguments{
\item{nutrients}{A character vector of nutrients of interest (e.g., "Iron", "Calcium", etc.)}

\item{isos}{A character vector of ISO3s of countries of interest (e.g., "USA")}

\item{sexes}{A character string of sexes of interest (i.e., "M", "F", or "MF")}

\item{ages}{A numeric vector of ages of interest (e.g., 10:50, 0:75, 55:90)}
}
\value{
A data frame with the parameters describing the requested subnational intake distributions
}
\description{
This function retrieves the subnational nutrient intake distributions modelled by Passarreli et al. (in prep).
}
\examples{
dists_all <- get_dists()
dists_usa <- get_dists(isos="USA")
dists_usa_iron <- get_dists(isos="USA", nutrients="Iron", sexes="MF")
dists_usa_iron_2040 <- get_dists(isos="USA", nutrients="Iron", sexes="MF", ages=20:40)
}
