% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate the kurtosis of a habitual intake distribution}
\usage{
kurtosis(shape = NULL, rate = NULL, meanlog, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\value{
The kurtosis of the distribution
}
\description{
This function calculates the kurtosis of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
kurtosis(shape=1, rate=0.5)
}
