% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_dist.R
\name{mean_dist}
\alias{mean_dist}
\title{Calculate the mean of a habitual intake distribution}
\usage{
mean_dist(shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\value{
The mean of the distribution
}
\description{
This function calculates the mean of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
mean_dist(shape=1, rate=0.5)
mean_dist(meanlog=3, sdlog=1)
}
