% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_dist.R
\name{median_dist}
\alias{median_dist}
\title{Calculate the median of a habitual intake distribution}
\usage{
median_dist(shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL, plot = F)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Plot (T/F)?}
}
\value{
The median of the distribution
}
\description{
This function calculates the median of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
median_dist(shape=1, rate=0.5)
median_dist(shape=1, rate=0.5, plot=T)
median_dist(meanlog=3, sdlog=1)
median_dist(meanlog=3, sdlog=1, plot=T)
}
