% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plots a log-normal or gamma distribution}
\usage{
plot_dist(shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\description{
This function plots a log-normal or gamma distribution.
}
\examples{
plot_dist(shape=2, rate=0.5)
plot_dist(meanlog=3, sdlog=1)
}
