% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dists.R
\name{plot_dists}
\alias{plot_dists}
\title{Plot subnational nutrient intake distributions}
\usage{
plot_dists(dists)
}
\arguments{
\item{dists}{A data frame of nutrient distributions}
}
\value{
A plot
dists <- dists_full %>% slice(1:3)
}
\description{
This function prepares a quick plot comparing subnational nutrient intake distributions. It is provided a dataframe generated from the get_dists() function.
}
