% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_dists.R
\name{shift_dist}
\alias{shift_dist}
\title{Shift the mean of a habitual intake distribution}
\usage{
shift_dist(
  shape = NULL,
  rate = NULL,
  meanlog = NULL,
  sdlog = NULL,
  by = NULL,
  to = NULL,
  plot = F
)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{by}{Amount to shift the current mean by}

\item{to}{New mean to shift the current mean to}

\item{plot}{Boolean (TRUE/FALSE) indicating whether to plot the current/shifted distributions}
}
\value{
A list with the parameters for the shifted distribution
}
\description{
This function shifts the mean of a habitual intake distribution and provides the new parameters used to describe the shifted distribution.
}
\examples{
# Shift gamma distribution by an amount (2 and -3)
shift_dist(shape=9.7, rate=1.25, by=2, plot=T)
shift_dist(shape=9.7, rate=1.25, by=-3, plot=T)

# Shift gamma distribution to a new mean (10 or 4)
shift_dist(shape=9.7, rate=1.25, to=10, plot=T)
shift_dist(shape=9.7, rate=1.25, to=4, plot=T)

# Shift log-normal distribution by an amount (10 and -5)
shift_dist(meanlog=3, sdlog=0.5, by=10, plot=T)
shift_dist(meanlog=3, sdlog=0.5, by=-5, plot=T)

# Shift log-normal distribution to a new mean (35 and 15)
shift_dist(meanlog=3, sdlog=0.5, to=35, plot=T)
shift_dist(meanlog=3, sdlog=0.5, to=15, plot=T)
}
