% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_req.R
\name{shift_req}
\alias{shift_req}
\title{Calculate intake mean shift required to achieve a specified prevalence}
\usage{
shift_req(
  ear,
  cv,
  target,
  shape = NULL,
  rate = NULL,
  meanlog = NULL,
  sdlog = NULL,
  plot = F
)
}
\arguments{
\item{ear}{Estimated Average Requirement (EAR)}

\item{cv}{Coefficient of variation (CV) of the EAR}

\item{target}{Target prevalence (0-100): 0 percent inadequate to 100 percent inadequate}

\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Boolean (TRUE/FALSE) indicating whether to plot the distributions relative to the EAR}
}
\value{
A list of the parameters for the intake distribution required to obtain the target prevalence
}
\description{
This function calculates the shift in a habitual intake distribution required to achieve a user-specified target prevalence of inadequate intakes.
}
\examples{
# Gamma-distribution examples
sev(ear=8.1, cv=0.1, shape=9.5, rate=1.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=25, shape=9.5, rate=1.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=10, shape=9.5, rate=1.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=5, shape=9.5, rate=1.3, plot=T)

# Lognormal-distribution examples
sev(ear=8.1, cv=0.1, meanlog=1.9, sdlog=0.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=25, meanlog=1.9, sdlog=0.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=10, meanlog=1.9, sdlog=0.3, plot=T)
shift_req(ear=8.1, cv=0.1, target=5, meanlog=1.9, sdlog=0.3, plot=T)
}
