% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Calculate the skewness of a habitual intake distribution}
\usage{
skewness(shape = NULL, rate = NULL, meanlog, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\value{
The skewness of the distribution
}
\description{
This function calculates the skewness of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
skewness(shape=1, rate=0.5)
skewness(meanlog=3, sdlog=1)
}
