% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Calculate the variance of a habitual intake distribution}
\usage{
variance(shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}
}
\value{
The variance of the distribution
}
\description{
This function calculates the variance of a habitual intake distribution. It handles gamma and log-normal distributions.
}
\examples{
variance(shape=1, rate=0.5)
variance(meanlog=3, sdlog=1)
}
