% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sev.R
\name{calc_sev}
\alias{calc_sev}
\title{Calculate summary exposure value (SEV, % deficiency)}
\usage{
calc_sev(
  ear,
  cv,
  shape = NULL,
  rate = NULL,
  meanlog = NULL,
  sdlog = NULL,
  plot = F
)
}
\arguments{
\item{ear}{Estimated Average Requirement (EAR)}

\item{cv}{Coefficient of variation (CV)}

\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Plot the distributions relative to the EAR? True/False}
}
\value{
The percent of a population deficient in a nutrient (SEV)
}
\description{
This function calculates the summary exposure value (SEV), or percent deficiency, for a group based on its Estimated Average Requirement and its habitual intake distribution.
}
