% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint.R
\name{cutpoint}
\alias{cutpoint}
\title{Calculate prevalence of inadequate intakes using the cutpoint method}
\usage{
cutpoint(intake_avg, intake_cv, ear)
}
\arguments{
\item{intake_avg}{Mean of the habitual intake distribution}

\item{intake_cv}{Coefficient of variation (CV) of the habitual intake distribution}

\item{ear}{Estimated Average Requirement (EAR)}
}
\value{
The percent of a population deficient in a nutrient (SEV)
}
\description{
This function calculates the prevalence of inadequate intakes using the EAR cutpoint method.
}
