% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dists.R
\name{generate_dists}
\alias{generate_dists}
\title{Generate subnational nutrient intake distributions}
\usage{
generate_dists(dists, perc = 0.9999)
}
\arguments{
\item{dists}{A data frame of nutrient distributions}

\item{perc}{The maximum percentile to draw to}
}
\value{
A dataframe
}
\description{
This function generates data for plotting subnational nutrient intake distributions. It is provided a dataframe generated from the get_dists() function.
}
\examples{
dists <- nutriR::get_dists(isos=c("USA", "BGD"), nutrients=c("Iron"), sexes="MF", ages=20:40)
dists_sim <- nutriR::generate_dists(dists)
}
