% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sev.R
\name{sev}
\alias{sev}
\title{Calculate summary exposure values}
\usage{
sev(ear, cv, shape = NULL, rate = NULL, meanlog = NULL, sdlog = NULL, plot = F)
}
\arguments{
\item{ear}{Estimated Average Requirement (EAR)}

\item{cv}{Coefficient of variation (CV)}

\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Plot the distributions relative to the EAR? True/False}
}
\value{
The percent of a population deficient in a nutrient (SEV)
}
\description{
This function calculates the summary exposure value (SEV), or percent deficiency, for a group based on its Estimated Average Requirement and its habitual intake distribution.
}
\examples{
sev(ear=8.1, cv=0.1, shape=9.5, rate=1.3, plot=T)
sev(ear=8.1, cv=0.1, meanlog=1.9, sdlog=0.3, plot=T)
}
