% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sev.R
\name{sev}
\alias{sev}
\title{Calculate summary exposure value}
\usage{
sev(
  ear,
  cv,
  ear_log = F,
  mean = NULL,
  sd = NULL,
  shape = NULL,
  rate = NULL,
  meanlog = NULL,
  sdlog = NULL,
  plot = F
)
}
\arguments{
\item{ear}{Estimated Average Requirement (EAR)}

\item{cv}{Coefficient of variation (CV) of the EAR}

\item{ear_log}{Boolean (TRUE/FALSE) indicating if the requirement distribution is lognormal. Default=F.}

\item{mean}{Mean for normal distribution}

\item{sd}{Standard deviation for normal distribution}

\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{plot}{Boolean (TRUE/FALSE) indicating whether to plot the distributions relative to the EAR}
}
\value{
The percent of a population with inadequate nutrient intakes (SEV)
}
\description{
This function calculates the summary exposure value (SEV), i.e., the percent of a population with inadequate nutrient intakes, for a group based on its Estimated Average Requirement (EAR) and its habitual intake distribution. It uses the probability method (NRC 1986). A CV of 0.25 is recommended for the EAR of Vitamin B12 and a CV of 0.10 is recommended for the EAR of all other nutrients (Renwick et al. 2004).
}
\examples{
sev(ear=8.1, cv=0.1, mean=8.5, sd=1.3, plot=T)
sev(ear=8.1, cv=0.1, shape=9.5, rate=1.3, plot=T)
sev(ear=8.1, cv=0.1, meanlog=1.9, sdlog=0.3, plot=T)
sev(ear=8.1, cv=0.44, ear_log=T, meanlog=1.84, sdlog=0.35, plot=T) # log-normal iron requirement
}
\references{
National Research Council (1986) Nutrient Adequacy: Assessment Using Food Consumption Surveys. Washington, DC: The National Academies Press. https://doi.org/10.17226/618.

Renwick AG, Flynn A, Fletcher RJ, Müller DJ, Tuijtelaars S, Verhagen H (2004) Risk-benefit analysis of micronutrients. Food and Chemical Toxicology 42(12): 1903-22. https://doi.org/10.1016/j.fct.2004.07.013
}
