% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_dists.R
\name{shift_dist}
\alias{shift_dist}
\title{Shift the mean of a habitual intake distribution}
\usage{
shift_dist(
  shape = NULL,
  rate = NULL,
  meanlog = NULL,
  sdlog = NULL,
  by = NULL,
  to = NULL,
  plot = F
)
}
\arguments{
\item{shape}{Shape parameter for gamma distribution}

\item{rate}{Rate parameter for gamma distribution}

\item{meanlog}{Mean parameter for gamma distribution}

\item{sdlog}{Standard deviation parameter for gamma distribution}

\item{by}{Amount to shift the current mean by}

\item{to}{New mean to shift the current mean to}

\item{plot}{Plot the current/shifted distributions? True/False}
}
\value{
A list with the parameters for the shifted distribution
}
\description{
This function shifts the mean of a habitual intake distribution and provides the new parameters used to describe the shifted distribution.
}
\examples{
# Shift distributions by an amount (2 and -3)
shift_dist(shape=9.7, rate=1.25, by=2, plot=T)
shift_dist(shape=9.7, rate=1.25, by=-3, plot=T)

# Shift distributions to a new mean (10 or 4)
shift_dist(shape=9.7, rate=1.25, to=10, plot=T)
shift_dist(shape=9.7, rate=1.25, to=4, plot=T)
}
